cpu-aarch64: Avoid hitting assert in GetCacheType()
With the advent of FEAT_MTE2, CTR_EL0 now has bits [37:32] defined as
TminLine. So this would result in relevant hardware tripping the
assertion in GetCacheType(), despite being otherwise normal behavior.

+++ src/aarch64/cpu-aarch64.h
@@ -320,7 +320,7 @@ class CPU {
 #undef VIXL_READ_ID_REG
 
   // Return the content of the cache type register.
-  static uint32_t GetCacheType();
+  static uint64_t GetCacheType();
 
   // I and D cache line size in bytes.
   static unsigned icache_line_size_;
